Move vec tests into trait
