+++ module-system/sov-modules-api/src/containers/accessory_vec.rs
@@ -105,125 +105,22 @@ where
 
 #[cfg(all(test, feature = "native"))]
 mod test {
-    use std::fmt::Debug;
 
     use sov_modules_core::WorkingSet;
     use sov_state::{DefaultStorageSpec, ProverStorage};
 
     use super::*;
+    use crate::containers::traits::vec_tests::Testable;
     use crate::default_context::DefaultContext;
 
-    enum TestCaseAction<T> {
-        Push(T),
-        Pop(T),
-        Last(T),
-        Set(usize, T),
-        SetAll(Vec<T>),
-        CheckLen(usize),
-        CheckContents(Vec<T>),
-        CheckContentsReverse(Vec<T>),
-        CheckGet(usize, Option<T>),
-        Clear,
-    }
-
-    fn test_cases() -> Vec<TestCaseAction<u32>> {
-        vec![
-            TestCaseAction::Push(1),
-            TestCaseAction::Push(2),
-            TestCaseAction::CheckContents(vec![1, 2]),
-            TestCaseAction::CheckLen(2),
-            TestCaseAction::Pop(2),
-            TestCaseAction::Set(0, 10),
-            TestCaseAction::CheckContents(vec![10]),
-            TestCaseAction::Push(8),
-            TestCaseAction::CheckContents(vec![10, 8]),
-            TestCaseAction::SetAll(vec![10]),
-            TestCaseAction::CheckContents(vec![10]),
-            TestCaseAction::CheckGet(1, None),
-            TestCaseAction::Set(0, u32::MAX),
-            TestCaseAction::Push(8),
-            TestCaseAction::Push(0),
-            TestCaseAction::CheckContents(vec![u32::MAX, 8, 0]),
-            TestCaseAction::SetAll(vec![11, 12]),
-            TestCaseAction::CheckContents(vec![11, 12]),
-            TestCaseAction::SetAll(vec![]),
-            TestCaseAction::CheckLen(0),
-            TestCaseAction::Push(42),
-            TestCaseAction::Push(1337),
-            TestCaseAction::Clear,
-            TestCaseAction::CheckContents(vec![]),
-            TestCaseAction::CheckGet(0, None),
-            TestCaseAction::SetAll(vec![1, 2, 3]),
-            TestCaseAction::CheckContents(vec![1, 2, 3]),
-            TestCaseAction::CheckContentsReverse(vec![3, 2, 1]),
-            TestCaseAction::Last(3),
-        ]
-    }
-
     #[test]
-    fn test_state_vec() {
+    fn test_accessory_state_vec() {
         let tmpdir = tempfile::tempdir().unwrap();
         let storage = ProverStorage::<DefaultStorageSpec>::with_path(tmpdir.path()).unwrap();
         let mut working_set: WorkingSet<DefaultContext> = WorkingSet::new(storage);
 
         let prefix = Prefix::new("test".as_bytes().to_vec());
         let state_vec = AccessoryStateVec::<u32>::new(prefix);
-
-        for test_case_action in test_cases() {
-            check_test_case_action(
-                &state_vec,
-                test_case_action,
-                &mut working_set.accessory_state(),
-            );
-        }
-    }
-
-    fn check_test_case_action<'ws, T, C>(
-        state_vec: &AccessoryStateVec<T>,
-        action: TestCaseAction<T>,
-        ws: &'ws mut AccessoryWorkingSet<'ws, C>,
-    ) where
-        BorshCodec: StateValueCodec<T> + StateValueCodec<usize>,
-        T: Eq + Debug,
-        C: Context,
-    {
-        match action {
-            TestCaseAction::CheckContents(expected) => {
-                let contents: Vec<T> = state_vec.iter(ws).collect();
-                assert_eq!(expected, contents);
-            }
-            TestCaseAction::CheckLen(expected) => {
-                let actual = state_vec.len(ws);
-                assert_eq!(actual, expected);
-            }
-            TestCaseAction::Pop(expected) => {
-                let actual = state_vec.pop(ws);
-                assert_eq!(actual, Some(expected));
-            }
-            TestCaseAction::Push(value) => {
-                state_vec.push(&value, ws);
-            }
-            TestCaseAction::Set(index, value) => {
-                state_vec.set(index, &value, ws).unwrap();
-            }
-            TestCaseAction::SetAll(values) => {
-                state_vec.set_all(values, ws);
-            }
-            TestCaseAction::CheckGet(index, expected) => {
-                let actual = state_vec.get(index, ws);
-                assert_eq!(actual, expected);
-            }
-            TestCaseAction::Clear => {
-                state_vec.clear(ws);
-            }
-            TestCaseAction::Last(expected) => {
-                let actual = state_vec.last(ws);
-                assert_eq!(actual, Some(expected));
-            }
-            TestCaseAction::CheckContentsReverse(expected) => {
-                let contents: Vec<T> = state_vec.iter(ws).rev().collect();
-                assert_eq!(expected, contents);
-            }
-        }
+        state_vec.run_tests(&mut working_set.accessory_state())
     }
 }
